% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{function1
Creates a boxplot of \code{varx} versus \code{vary} with ggplot.}
\usage{
function1(df, varx, vary, title, xlab, ylab, legend)
}
\arguments{
\item{df}{dataframe name}

\item{varx}{string name for variable x in the dataframe}

\item{vary}{string name for variable y in the dataframe}

\item{title}{text data for title of the boxplot}

\item{xlab}{text data for the x-axis for the boxplot}

\item{ylab}{text data for the y-axis of the boxplot}

\item{legend}{text data for the legend title of the boxplot}
}
\value{
a boxplot of \code{varx} versus \code{vary}, with title
x and y labels, and legend title.
}
\description{
function1
Creates a boxplot of \code{varx} versus \code{vary} with ggplot.
}
\examples{
function1(df = drg,
         varx = "DRG.Definition",
         vary = "Average.Total.Payments",
         title = "title name",
         xlab = "x label",
         ylab = "y label",
         legend = "legend title")


}
