% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{function2}
\alias{function2}
\title{function2
groups data by \code{group}, followed by a specific summary statistic for a
selected variable \code{value} within the grouped data.}
\usage{
function2(data, group, fun, value)
}
\arguments{
\item{data}{is the dataframe name}

\item{group}{string name for variable to group by.}

\item{fun}{name of the function to be used. Example: mean, median, or std.}

\item{value}{string name for variable to use the summary statistic function on.}
}
\value{
a table with two columns 1) \code{group}, and 2) a column with the name \code{statistic} that shows the value of the statistic.
}
\description{
function2
groups data by \code{group}, followed by a specific summary statistic for a
selected variable \code{value} within the grouped data.
}
\examples{
function2(data = drg,
         group = DRG.Definition,
         fun = mean,
         value = Average.Medicare.Payments)



}
